Initial documentation

+++ wrfutilV3.1/bin/main_WPS.sh
@@ -27,7 +27,15 @@ source $BASEDIR/conf/machine.conf
 ####################################
 #Calculamos la cantidad de pasos
 ####################################
-PASOS_RESTANTES=$((($(date -d "$FECHA_FIN" +%s) - $(date -d "$FECHA_INI" +%s))/$ANALISIS_FREC ))
+if [ $EXPTYPE == 'assimilation' ] ; then 
+   PASOS_RESTANTES=$((($(date -d "$FECHA_FIN" +%s) - $(date -d "$FECHA_INI" +%s)) / $ANALISIS_FREC ))
+elif [ $EXPTYPE == 'forecast' ]  ; then
+   PASOS_RESTANTES=$((( ($(date -d "$FECHA_FIN" +%s) - $(date -d "$FECHA_INI" +%s)) - $FORECAST_LEAD_TIME ) / $FORECAST_INI_FREQ + 1 ))
+else
+   echo "Error: Not recognized experiment type."
+   exit
+fi
+
 PASOS_RESTANTES=$((10#$PASOS_RESTANTES-10#$PASO))
 
 echo "El experimento abarca desde $FECHA_INI hasta $FECHA_FIN"
@@ -38,21 +46,27 @@ rm -f $PROCSDIR/*_ENDOK
 ###### 1st assimilation cycle only
 echo " Step | TimeStamp" > $LOGDIR/cycles.log
 if [ $RUN_WPS -eq 1 ] ; then 
-   echo "Corriendo el WPS"
-   time $BASEDIR/bin/correr_WPS.sh > $LOGDIR/log_wps_${PASO}.log
+   echo "Running WPS"
+   time $BASEDIR/bin/run_WPS.sh > $LOGDIR/log_wps_${PASO}.log
 fi
 if [[ $BDY_PERT -eq 1 && $RUN_BDY_PERT -eq 1 ]] ; then
-   echo "Vamos a perturbar los met_em"
+   if [ $MACHINE == 'FUGAKU' ] ; then 
+      echo "Warning: met_em perturbation can not be executed at the computer nodes of FUGAKU"
+      echo "Please execute main_Pert.sh in the login nodes or in the FUGAKU pre/post nodes"
+      echo "      -----------------------------------------------------------             "
+      exit
+   fi      
+   echo "We will perturb met_em files"
    rm -fr $HISTDIR/WPS/met_em
-   time $BASEDIR/bin/correr_Pert.sh > $LOGDIR/pert_met_em_${PASO}.log   2>&1
+   time $BASEDIR/bin/run_Pert.sh > $LOGDIR/pert_met_em_${PASO}.log   2>&1
 elif [ $BDY_PERT -eq 0 ] ; then
    echo "Linking met_em directory"
    rm -fr $HISTDIR/WPS/met_em
    time ln -sf $HISTDIR/WPS/met_em_ori $HISTDIR/WPS/met_em > $LOGDIR/pert_met_em_${PASO}.log  2>&1
 fi
 
-echo "Saliendo. . .  . hasta la proxima!"
-echo "Termiamos de correr a:"$(date )
+echo "Successfully finish running WPS"
+echo "We finish running @"$(date )
 exit 0
